requirements.txt
astroid==2.4.2
bandit==1.6.2
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
colorama==0.4.4
decorator==4.4.2
dnspython==2.0.0
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
html5lib==1.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
lxml==4.6.1
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.4
pbr==5.5.1
pylint==2.6.0
python-dateutil==2.8.1
python-nmap==0.6.1
python-whois==0.7.3
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
shodan==1.24.0
six==1.15.0
smmap==3.0.4
soupsieve==2.0.1
stevedore==3.2.2
toml==0.10.2
urllib3==1.25.11
validators==0.18.1
vulners==1.5.9
webencodings==0.5.1
wrapt==1.12.1
XlsxWriter==1.3.7
